Fix and improve data ingestion
the `ingest` function inserted readings with the wrong date because it
always appended `CEST` for the timezone which caused PostgreSQL (or
Psycopg2?) to interpret the time as UTC+2 regardless of the actual
date (in winter time, i.e. CET, it is actually UTC+1).

Now with the help of pytz, the correct time zone information is passed
to PostgreSQL.

The `readings` command now wraps all SQL statements in a single
transaction by passing `autocommit=False` to the `connect` method.
This changes the behaviour in a major way in that a single failure in
any SQL statement will cause no changes to the DB. I find this a much
safer way to handle errors and leave the DB in a consistent state.

Additionally, this commit changes `ingest` in the following ways,
making it run much faster than before:

1. It only inserts values for dates that aren't stored in the `gas`
   table, yet by querying the most recent entry and skipping over all
   lines in the input CSV that are older (or equal) to that entry's
   date.
2. It to use a multi-value insert statement which makes the function
   perform much faster than when it used a single insert statement per
   row.